feat: allow new insight type choice from sidebar
