+++ frontend/src/layout/navigation/SideBar/SideBar.tsx
@@ -17,7 +17,6 @@ import {
     IconOpenInApp,
     IconPerson,
     IconPinOutline,
-    IconPlus,
     IconRecording,
     IconSettings,
     IconTools,
@@ -49,6 +48,7 @@ import Typography from 'antd/lib/typography'
 import { Spinner } from 'lib/lemon-ui/Spinner/Spinner'
 import { DebugNotice } from 'lib/components/DebugNotice'
 import ActivationSidebar from 'lib/components/ActivationSidebar/ActivationSidebar'
+import { overlayForNewInsightMenu } from 'scenes/saved-insights/newInsightsMenu'
 
 function Pages(): JSX.Element {
     const { currentOrganization } = useValues(organizationLogic)
@@ -64,6 +64,11 @@ function Pages(): JSX.Element {
 
     const [arePinnedDashboardsShown, setArePinnedDashboardsShown] = useState(false)
     const [isToolbarLaunchShown, setIsToolbarLaunchShown] = useState(false)
+    const [isNewInsightMenuShown, setIsNewInsightMenuShown] = useState(false)
+
+    const isUsingDataExplorationQueries =
+        !!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_INSIGHTS] &&
+        !!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_QUERY_TAB]
 
     return (
         <ul>
@@ -152,11 +157,25 @@ function Pages(): JSX.Element {
                         identifier={Scene.SavedInsights}
                         to={urls.savedInsights()}
                         sideAction={{
-                            icon: <IconPlus />,
-                            to: urls.insightNew(),
-                            tooltip: 'New insight',
                             identifier: Scene.Insight,
-                            onClick: hideSideBarMobile,
+                            tooltip: 'Choose an insight type to create',
+                            onClick: () => setIsNewInsightMenuShown((state) => !state),
+                            dropdown: {
+                                visible: isNewInsightMenuShown,
+                                onClickOutside: () => setIsNewInsightMenuShown(false),
+                                onClickInside: hideSideBarMobile,
+                                overlay: (
+                                    <div className="SideBar__side-actions" data-attr="sidebar-new-insights">
+                                        <h5>New Insight</h5>
+                                        <LemonDivider />
+                                        {overlayForNewInsightMenu(
+                                            'sidebar-new-insights-overlay',
+                                            isUsingDataExplorationQueries,
+                                            () => setIsNewInsightMenuShown(false)
+                                        )}
+                                    </div>
+                                ),
+                            },
                         }}
                     />
                     <PageButton
