+++ frontend/src/scenes/saved-insights/SavedInsights.tsx
@@ -33,7 +33,6 @@ import {
 import { SceneExport } from 'scenes/sceneTypes'
 import { TZLabel } from 'lib/components/TZLabel'
 import { urls } from 'scenes/urls'
-import { eventUsageLogic } from 'lib/utils/eventUsageLogic'
 import { LemonTable, LemonTableColumn, LemonTableColumns } from 'lib/lemon-ui/LemonTable'
 import { LemonDivider } from 'lib/lemon-ui/LemonDivider'
 import { More } from 'lib/lemon-ui/LemonButton/More'
@@ -56,7 +55,7 @@ import { LemonTabs } from 'lib/lemon-ui/LemonTabs'
 import { featureFlagLogic } from 'lib/logic/featureFlagLogic'
 import { FEATURE_FLAGS } from 'lib/constants'
 import { isInsightVizNode } from '~/queries/utils'
-import { examples } from '~/queries/examples'
+import { overlayForNewInsightMenu } from 'scenes/saved-insights/newInsightsMenu'
 
 interface NewInsightButtonProps {
     dataAttr: string
@@ -257,17 +256,6 @@ export const scene: SceneExport = {
     logic: savedInsightsLogic,
 }
 
-const insightTypeURL: Record<InsightType, string> = {
-    TRENDS: urls.insightNew({ insight: InsightType.TRENDS }),
-    STICKINESS: urls.insightNew({ insight: InsightType.STICKINESS }),
-    LIFECYCLE: urls.insightNew({ insight: InsightType.LIFECYCLE }),
-    FUNNELS: urls.insightNew({ insight: InsightType.FUNNELS }),
-    RETENTION: urls.insightNew({ insight: InsightType.RETENTION }),
-    PATHS: urls.insightNew({ insight: InsightType.PATHS }),
-    QUERY: urls.insightNew(undefined, undefined, JSON.stringify(examples.EventsTableFull)),
-    SQL: urls.insightNew(undefined, undefined, JSON.stringify(examples.HogQLTable)),
-}
-
 export function InsightIcon({ insight }: { insight: InsightModel }): JSX.Element | null {
     let insightType = insight?.filters?.insight || InsightType.TRENDS
     if (!!insight.query && !isInsightVizNode(insight.query)) {
@@ -283,13 +271,6 @@ export function InsightIcon({ insight }: { insight: InsightModel }): JSX.Element
 export function NewInsightButton({ dataAttr }: NewInsightButtonProps): JSX.Element {
     const { featureFlags } = useValues(featureFlagLogic)
 
-    let menuEntries = Object.entries(INSIGHT_TYPES_METADATA)
-    if (!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_QUERY_TAB]) {
-        menuEntries = menuEntries.filter(
-            ([insightType]) => insightType !== InsightType.QUERY && insightType !== InsightType.SQL
-        )
-    }
-
     return (
         <LemonButtonWithSideAction
             type="primary"
@@ -299,31 +280,9 @@ export function NewInsightButton({ dataAttr }: NewInsightButtonProps): JSX.Eleme
                     placement: 'bottom-end',
                     className: 'new-insight-overlay',
                     actionable: true,
-                    overlay: menuEntries.map(
-                        ([listedInsightType, listedInsightTypeMetadata]) =>
-                            listedInsightTypeMetadata.inMenu && (
-                                <LemonButton
-                                    key={listedInsightType}
-                                    status="stealth"
-                                    icon={
-                                        listedInsightTypeMetadata.icon && (
-                                            <listedInsightTypeMetadata.icon color="var(--muted-alt)" noBackground />
-                                        )
-                                    }
-                                    to={insightTypeURL[listedInsightType as InsightType]}
-                                    data-attr={dataAttr}
-                                    data-attr-insight-type={listedInsightType}
-                                    onClick={() => {
-                                        eventUsageLogic.actions.reportSavedInsightNewInsightClicked(listedInsightType)
-                                    }}
-                                    fullWidth
-                                >
-                                    <div className="text-default flex flex-col text-sm py-1">
-                                        <strong>{listedInsightTypeMetadata.name}</strong>
-                                        <span className="text-xs">{listedInsightTypeMetadata.description}</span>
-                                    </div>
-                                </LemonButton>
-                            )
+                    overlay: overlayForNewInsightMenu(
+                        dataAttr,
+                        !!featureFlags[FEATURE_FLAGS.DATA_EXPLORATION_QUERY_TAB]
                     ),
                 },
                 'data-attr': 'saved-insights-new-insight-dropdown',
